Redo moving of CodeAndName type
