--- frontend/src/domain/entities/types.ts
@@ -1,4 +1,0 @@
-export type CodeAndName = {
-  code: string
-  name: string
-}
