+++ frontend/src/features/MapButtons/LayersSidebar/AdministrativeZones/AdministrativeZonesGroup.tsx
@@ -7,8 +7,7 @@ import { useMainAppSelector } from '../../../../hooks/useMainAppSelector'
 import { theme } from '../../../../ui/theme'
 import { ChevronIcon } from '../../../commonStyles/icons/ChevronIcon.style'
 
-import type { ShowableLayer } from '../../../../domain/entities/layers/types'
-import type { CodeAndName } from '../../../../domain/entities/types'
+import type { CodeAndName, ShowableLayer } from '../../../../domain/entities/layers/types'
 
 type AdministrativeZonesGroupType = {
   group: CodeAndName
